Just making blur use uint8 gotten from convolution

+++ blur.m
@@ -2,5 +2,5 @@ function x = blur(img)
   w = [1/16, 2/16, 1/16;
        2/16, 4/16, 2/16;
        1/16, 2/16, 1/16];
-  x = convolution(img, w);
+  x = uint8(convolution(img, w));
 endfunction
